helpers converted, typescript base
