+++ tsconfig.json
@@ -1,0 +1,16 @@
+{
+  "compilerOptions": {
+    "target": "es2016",
+    "module": "commonjs",
+    "allowJs": true,
+    "checkJs": false,
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "strict": false, 
+    "alwaysStrict": true,
+    "skipLibCheck": true
+  },
+  "include": [
+    "./src/*.ts"
+  ]
+}
